requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
APScheduler==3.11.2
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
pandas-datareader